fix: cpr_bridge_http — int cast for cpr::Timeout (silence -Wnarrowing; telegram precedent)

+++ src/bridge/cpr_bridge_http.cpp
@@ -11,7 +11,7 @@ std::pair<int, std::string> CprBridgeHttp::post_json(const std::string& url,
     cpr::Response r = cpr::Post(
         cpr::Url{url}, cpr::Body{body},
         cpr::Header{{"Content-Type", "application/json"}, {"X-Hades-Bridge", secret}},
-        cpr::Timeout{static_cast<long>(timeout_s * 1000)}, cpr::Redirect{false});
+        cpr::Timeout{static_cast<int>(timeout_s * 1000)}, cpr::Redirect{false});
     return {static_cast<int>(r.status_code), r.text};
   } catch (const std::exception&) {
     return {0, ""};   // transport failure — the caller degrades (never throws)
